perf: switch compression to fast level and eliminate compress_tmp intermediate buffer
- Compression::best() -> Compression::fast() for 3-5x CPU reduction
- Remove compress_tmp Vec, write compressed output directly to dst BytesMut
- Eliminates one O(n) memcpy per frame on the encode hot path
- Fix clippy warnings: explicit_auto_deref, unnecessary_cast

+++ rfrp_client/src/lib.rs
@@ -61,7 +61,7 @@ pub async fn rfrp_client(config: Arc<ConfigInfo>) {
 
 /// Calculate reconnection delay with exponential backoff, capped at MAX_RECONNECT_DELAY.
 fn calc_reconnect_delay(attempt: u32) -> Duration {
-    let seconds = (INITIAL_RECONNECT_DELAY as u64)
+    let seconds = INITIAL_RECONNECT_DELAY
         .saturating_mul(2u64.saturating_pow(attempt.saturating_sub(1)))
         .min(MAX_RECONNECT_DELAY);
     Duration::from_secs(seconds)
